reformats listarMorosos
